Added RETAIN_DAY_OF_WEEK option

+++ ebs-snapshot.sh
@@ -20,10 +20,13 @@ set -o pipefail
 # - Take a snapshot of each attached volume
 # - The script will then delete all associated snapshots taken by the script that are older than 7 days
 #
-# DISCLAIMER: This script deletes snapshots (though only the ones that it creates). 
+# DISCLAIMER: This script deletes snapshots (though only the ones that it creates).
 # Make sure that you understand how the script works. No responsibility accepted in event of accidental data loss.
 #
-
+# ENV VARS:
+#   AWS_CONFIG_FILE - The location of your AWS config file.
+#   RETAIN_DAY_OF_WEEK - (Optional) An integer indicating a day of the week for which you would like to
+#     retain backups indefinitely. 1 is Monday, 2 is Tuesday, etc.
 
 ## Variable Declartions ##
 
@@ -80,7 +83,7 @@ snapshot_volumes() {
 
 		snapshot_id=$(aws ec2 create-snapshot --region $region --output=text --description $snapshot_description --volume-id $volume_id --query SnapshotId)
 		log "New snapshot is $snapshot_id"
-	 
+
 		# Add a "CreatedBy:AutomatedBackup" tag to the resulting snapshot.
 		# Why? Because we only want to purge snapshots taken by the script later, and not delete snapshots manually taken.
 		aws ec2 create-tags --region $region --resource $snapshot_id --tags Key=CreatedBy,Value=AutomatedBackup
@@ -99,14 +102,16 @@ cleanup_snapshots() {
 			snapshot_description=$(aws ec2 describe-snapshots --snapshot-id $snapshot --region $region --query Snapshots[].Description)
 
 			if (( $snapshot_date_in_seconds <= $retention_date_in_seconds )); then
-				log "DELETING snapshot $snapshot. Description: $snapshot_description ..."
-				aws ec2 delete-snapshot --region $region --snapshot-id $snapshot
+				if [[ $(date --date=$snapshot_date +%u) == $RETAIN_DAY_OF_WEEK ]]; then
+					log "DELETING snapshot $snapshot. Description: $snapshot_description ..."
+					aws ec2 delete-snapshot --region $region --snapshot-id $snapshot
+				fi
 			else
 				log "Not deleting snapshot $snapshot. Description: $snapshot_description ..."
 			fi
 		done
 	done
-}	
+}
 
 
 ## SCRIPT COMMANDS ##
